backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djoser==2.2.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
Pillow==10.2.0
gunicorn==21.2.0
#setuptools==69.0.3
django-cors-headers==4.3.1
python-dotenv==1.0.0
django-filter==23.5